feat: move Credential functions into Request and remove Credential file
