--- packages/core/src/credential/index.ts
@@ -1,8 +1,0 @@
-/**
- * Copyright (c) 2018-2022, BOTLabs GmbH.
- *
- * This source code is licensed under the BSD 4-Clause "Original" license
- * found in the LICENSE file in the root directory of this source tree.
- */
-
-export * from './Credential.js'
