+++ packages/core/src/index.ts
@@ -12,13 +12,12 @@
 import '@polkadot/api-augment'
 
 export * as Attestation from './attestation/index.js'
-export * as Credential from './credential/index.js'
+export * as Credential from './requestforattestation/index.js'
 export { Balance, BalanceUtils } from './balance/index.js'
 export * as Claim from './claim/index.js'
 export * as CType from './ctype/index.js'
 export { DelegationNode, DelegationNodeUtils } from './delegation/index.js'
 export * as Quote from './quote/index.js'
-export * as RequestForAttestation from './requestforattestation/index.js'
 
 export { connect, disconnect, config, init } from './kilt/index.js'
 export { SDKErrors } from '@kiltprotocol/utils'
